+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   - Upgraded Ruby SDK version to 2.0.0
   - Fixed add/create, delete/remove for resources
   - Added support to Enclosure group actions
+  - Added support for Enclosure actions
 
 ### 0.1.1
   - Fixed Ruby SDK version to 1.0.0